requirements.txt
# Claude Agent SDK
claude-agent-sdk>=0.1.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
httpx>=0.25.0

# Windows COM API
pywin32>=306

# Screen Capture & GUI Automation
mss>=9.0.0
pillow>=10.0.0
pyautogui>=0.9.54

# Data Validation
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
